+++ src/utils/container.ts
@@ -1,6 +1,7 @@
 import _isInsideContainer from 'is-inside-container'
 
 export const isInsideContainer = () =>
-  isOfficialDockerImage() || (_isInsideContainer() && !process.env.MARP_TEST_CI)
+  isOfficialContainerImage() ||
+  (_isInsideContainer() && !process.env.MARP_TEST_CI)
 
-export const isOfficialDockerImage = () => !!process.env.MARP_USER
+export const isOfficialContainerImage = () => !!process.env.MARP_USER